+++ .cargo/config.toml
@@ -1,83 +1,92 @@
-## Project Runners ##
-# Uncomment only one of the runners
-# If you want to use the defmt logger, uncomment the defmt runner
+# Cargo configuration for esp-csi-cli-rs
+#
+# This file documents the available Cargo aliases and target-specific settings
+# used when building and flashing ESP targets. It is included in docs.rs so
+# users can quickly see common commands and tooling expectations.
 
-## Default: ESP32-C3 Runner ##
-[target.riscv32imc-unknown-none-elf]
-# Run with defmt logging enabled
+[alias]
+# Convenience aliases for running examples on supported chips.
+# Example: `cargo esp32 --example wifi_station`
+esp32 = "run --release --features=esp32   --target=xtensa-esp32-none-elf"
+esp32c3 = "run --release --features=esp32c3 --target=riscv32imc-unknown-none-elf"
+esp32c6 = "run --release --features=esp32c6 --target=riscv32imac-unknown-none-elf"
+esp32c2 = "run --release --features=esp32c2 --target=riscv32imc-unknown-none-elf"
+esp32s3 = "run --release --features=esp32s3 --target=xtensa-esp32s3-none-elf"
+
+# Convenience aliases for building release binaries without running.
+esp32-build = "build --release --features=esp32   --target=xtensa-esp32-none-elf"
+esp32c3-build = "build --release --features=esp32c3 --target=riscv32imc-unknown-none-elf"
+esp32c6-build = "build --release --features=esp32c6 --target=riscv32imac-unknown-none-elf"
+esp32c2-build = "build --release --features=esp32c2 --target=riscv32imc-unknown-none-elf"
+esp32s3-build = "build --release --features=esp32s3 --target=xtensa-esp32s3-none-elf"
+
+[target.'cfg(target_arch = "riscv32")']
+# Default runner and linker flags for all RISC-V ESP targets.
 # runner = "espflash flash --monitor --log-format defmt"
-# Run without defmt logging enabled
 runner = "espflash flash --monitor"
 rustflags = [
   "-C",
   "link-arg=-Tlinkall.x",
   # Uncomment the following two lines if you want to use defmt
-  "-C",
-  "link-arg=-Tdefmt.x",
+  # "-C", 
+  # "link-arg=-Tdefmt.x",
 ]
 
+[target.'cfg(target_arch = "xtensa")']
+# Default runner and linker flags for all Xtensa ESP targets.
+# Uncomment the following line if you want to use defmt
+# runner = "espflash flash --monitor --log-format defmt"
+runner = "espflash flash --monitor"
+rustflags = [
+  # GNU LD
+  "-C",
+  "link-arg=-Wl,-Tlinkall.x",
+  "-C",
+  "link-arg=-nostartfiles",
+  # Uncomment the following two lines if you want to use defmt
+  # "-C",
+  # "link-arg=-Tdefmt.x",
 
-# # ESP32-C6 Runner ##
-# [target.riscv32imac-unknown-none-elf]
-# # Run with defmt logging enabled
-# # runner = "espflash flash --monitor --log-format defmt"
-# # Run without defmt logging enabled
-# runner = "espflash flash --monitor"
-# rustflags = [
-#   "-C",
-#   "link-arg=-Tlinkall.x",
-#   # Uncomment the following two lines if you want to use defmt
-#   # "-C",
-#   # "link-arg=-Tdefmt.x",
-# ]
+  # LLD
+  # "-C", "link-arg=-Tlinkall.x",
+  # "-C", "linker=rust-lld",
+]
 
-# # ESP32 Runner ##
-# [target.xtensa-esp32-none-elf]
-# # runner = "espflash flash --monitor --log-format defmt"
-# runner = "espflash flash --monitor"
-# rustflags = [
-#   # GNU LD
-#   "-C",
-#   "link-arg=-Wl,-Tlinkall.x",
-#   "-C",
-#   "link-arg=-nostartfiles",
-#   # Uncomment the following two lines if you want to use defmt
-#   "-C",
-#   "link-arg=-Tdefmt.x",
-# ]
+[target.riscv32imc-unknown-none-elf]
+# Per-target overrides for RISC-V IMC devices (e.g., ESP32-C2/C3).
+# Uncomment the following line if you want to use defmt
+# runner = "espflash flash --monitor --log-format defmt"
+# Comment the following line if you want to use defmt
+runner = "espflash flash --monitor"
 
-# # ESP32-S3 Runner ##
-# [target.xtensa-esp32s3-none-elf]
-# # runner = "espflash flash --monitor --log-format defmt"
-# runner = "espflash flash --monitor"
-# rustflags = [
-#   # GNU LD
-#   "-C",
-#   "link-arg=-Wl,-Tlinkall.x",
-#   "-C",
-#   "link-arg=-nostartfiles",
-#   # Uncomment the following two lines if you want to use defmt
-#   "-C",
-#   "link-arg=-Tdefmt.x",
-# ]
+[target.riscv32imac-unknown-none-elf]
+# Per-target overrides for RISC-V IMAC devices (e.g., ESP32-C6).
+# Uncomment the following line if you want to use defmt
+# runner = "espflash flash --monitor --log-format defmt"
+# Comment the following line if you want to use defmt
+runner = "espflash flash --monitor"
+
+[target.xtensa-esp32s3-none-elf]
+# Per-target overrides for Xtensa ESP32-S3 devices.
+# Uncomment the following line if you want to use defmt
+# runner = "espflash flash --monitor --log-format defmt"
+# Comment the following line if you want to use defmt
+runner = "espflash flash --monitor"
+
+[target.xtensa-esp32-none-elf]
+# Per-target overrides for Xtensa ESP32 devices.
+# Uncomment the following line if you want to use defmt
+# runner = "espflash flash --monitor --log-format defmt"
+# Comment the following line if you want to use defmt
+runner = "espflash flash --monitor"
 
 
 [env]
+# Default log level for ESP-IDF logging output.
 ESP_LOG = "INFO"
-
-## Build Targets
-# Uncomment only one of the build targets
-# NOTE: Beware of whitespaces between lines
-
-[build]
-# Default: ESP32-C3 Build Target
-target = "riscv32imc-unknown-none-elf"
-# ESP32-C6 and ESP32-H2 Build Target
-# target = "riscv32imac-unknown-none-elf"
-# ESP32-S3 Build Target
-# target = "xtensa-esp32s3-none-elf"
-# ESP32 Build Target
-# target = "xtensa-esp32-none-elf"
+# Uncomment the following line if you want to use defmt
+# DEFMT_LOG = "INFO"
 
 [unstable]
+# Build core/alloc for no_std targets.
 build-std = ["alloc", "core"]
